Create load_json_inputs.jl
Separate JSON loading and object loading.

Simplified eager loading control flow and recursion. Also tried to make it more explicit so it is easier to follow.

Fixedbug in eager loading. The issue arose when we were expanding global -> instance data. When eager loading, we need to deepcopy the global data to avoid one instance changing data for all the instances, including when loading and converting inputs to objects.