Commodity of gasoline with exogenous demand and end use transformations
Model exogenous gasoline demand use with transformation to account for CO2 emissions from gasoline use. Similar to modeling of exogenous NG demand with a demand node. A transformation of gasoline end use is created with start vertex from the main gasoline node (currently has prices for conventional fossil gasoline but will act as a blending node when synthetic and bio liquid fuels are modeled), and end vertex to the CO2 sink, and a gasoline demand node.