Commodity of diesel with exogenous demand and end use transformations
Model exogenous diesel demand use with transformation to account for CO2 emissions from gasoline use. Similar to modeling of exogenous NG, gasoline, jet fuel demand with a demand node. A transformation of diesel end use is created with start vertex from the main diesel node (currently has prices for conventional fossil diesel but will act as a blending node when synthetic and bio liquid fuels are modeled), and end vertex to the CO2 sink, and a diesel demand node.